v20: Define core OCPP 2.0 messages structure
Only three messages are required to be considered OCPP 2.0:
ChangeAvailability, GetVariables and SetVariables.

It's not much but now ocpp-codec can be used to write an OCPP 2.0
server. However, both GetVariableResultType and SetVariableDataType
`attributeValue` should have a dynamic max length. That's not
implemented yet, the max length is the default value: 1000.